Removing lazy keyword from atomValidator declaration.  This prevents the test suite from taking an inordinate long time due to an error in the schemas.
When the Validator initializes with an exception, the lazy declaration for the atomValidator val causes the Validator to be reinitialized for each access to the val (i.e., each test) as opposed to once.  Currently there are ~68 tests.  Here's a guy talking about this lazy behavior:   https://twitter.com/viktorklang/status/104483846002704384.  This matches my observations.
Furthermore, there is a memory leak somewhere which causes the initialization to slow down.  By 20 tests in the time really starts ramping up:  1, 2, 27, 43, 48, 91, 778 seconds. I assume this is associated with initializing the validator.  A preliminary look didn't show much more than  "there are too many strings".  The strings look like to be related to status messages coming from the XML parser.  Strings & char[] are accounting for ~20% of the heap.