Generate Cloud Feeds Product Schemas reference appendix
src/docbkx/product_schemas_appendix.xml is a manifest file that lists
what product schema files should be included, in what order, and
using what sample message file. It also include text describing each
product schema as well as the whole appendix.

src/docbkx/xslts/productSchema2DocBook.xsl converts the manifest
to DocBook as a pre-processing step.

I've added a plugin to the pom to run the xslt and an xi:include
to the feed-devguide.xml to pull in the generated appendix.