CF-1038 - Implement feeds-specific RBAC for publishing to terminator feeds
Enables the ability to require product feeds to require specific roles to POST to the feed.  Default role is the current
cloudfeeds:service-admin.

To enable a custom role, the attribulte writeRole attribute is set in the productSchema node of of the product schema.

Roles can also be tested by adding the writeRole attribute to the atom node in the message samples.